requirements.txt
flask
anthropic
python-dotenv
pymupdf
sentence-transformers
numpy
gunicorn
